Design and Login Before Choose Menu

+++ src/component/Breakfast/BreakFastFish.js
@@ -9,31 +9,43 @@ export default class BreakFastFish extends Component {
     menu_price: 40,
   };
 
-  sentOrder = (e) => {
+  sentOrder = async (e) => {
     e.preventDefault();
+    let token = localStorage.getItem("token");
+    if (token) {
+      if (token !== null) {
+        this.setState({ check: "login" });
+        var oldItems = JSON.parse(localStorage.getItem("order")) || [];
+        let uid = localStorage.getItem("user_id");
+        console.log("sent !");
 
-    var oldItems = JSON.parse(localStorage.getItem("order")) || [];
-    let uid = localStorage.getItem("user_id");
-    console.log("sent !");
+        const newData = {
+          menu_name: this.state.menu_name,
+          menu_value: 1,
+          menu_price: this.state.menu_price,
+          user_id: uid,
+        };
 
-    const newData = {
-      menu_name: this.state.menu_name,
-      menu_value: 1,
-      menu_price: this.state.menu_price,
-      user_id: uid,
-    };
-    oldItems.push(newData);
-    localStorage.setItem("order", JSON.stringify(oldItems));
-    console.log(" data : ", newData);
-    this.setState({ menu_value: 0 });
-    this.setState({ remark: "" });
-  };
+        const isHaveMenuAlready = oldItems.find(
+          (menu) => menu.menu_name === this.state.menu_name
+        );
 
-  handleInputChange = (e) => {
-    const { name, value } = e.target;
-    this.setState({ [name]: value });
-    this.setState({ please: "" });
-    console.log({ [name]: value });
+        if (!isHaveMenuAlready) {
+          oldItems.push(newData);
+          localStorage.setItem("order", JSON.stringify(oldItems));
+          console.log(" data : ", newData);
+          this.setState({ menu_value: 0 });
+          this.setState({ remark: "" });
+          window.location.reload();
+        }
+      }
+      let res = await axios.post(`http://localhost:3001/api/users/me`, {
+        token,
+      });
+      this.setState({ user: res.data });
+    } else {
+      window.location.href = "/login";
+    }
   };
   render() {
     return (
